Automatically Merged using SST Master Branch Merger

+++ sstmac/hardware/pisces/pisces_packetizer.cc
@@ -169,7 +169,7 @@ pisces_packetizer::spaceToSend(int vn, int num_bits)
 }
 
 void
-pisces_packetizer::inject(int vn, long bytes, long byte_offset, message* msg)
+pisces_packetizer::inject(int vn, uint32_t bytes, uint64_t byte_offset, message* msg)
 {
   bool is_tail = (byte_offset + bytes) == msg->byte_length();
   //only carry the payload if you're the tail packet
